+++ screens/Authentication/Signin.tsx
@@ -1,0 +1,36 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+import { useAuthentication } from "../utils/hooks/useAuthentication";
+import { Button } from "react-native-elements";
+import { getAuth, signOut } from "firebase/auth";
+
+const auth = getAuth();
+
+const SigninScreen = ({ navigation }: any) => {
+  const { user } = useAuthentication();
+
+  return (
+    <View style={styles.container}>
+      <Text>Sign in</Text>
+      <Button
+        title="Back"
+        style={styles.button}
+        onPress={() => navigation.navigate("Home")}
+      />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  button: {
+    marginTop: 10,
+  },
+});
+
+export default SigninScreen;
